+++ scout/server/blueprints/omics_variants/__init__.py
@@ -1,0 +1,1 @@
+from .views import omics_variants_bp