+++ scout/server/blueprints/variants/controllers.py
@@ -1520,7 +1520,7 @@ def populate_filters_form(store, institute_obj, case_obj, user_obj, category, re
             }
         )
         clinical_filter = MultiDict(clinical_filter_dict)
-    elif category in ("sv", "cancer", "cancer_sv", "mei"):
+    elif category in ("sv", "cancer", "cancer_sv", "mei", "outlier"):
         clinical_filter_dict = FiltersFormClass.clinical_filter_base
         clinical_filter_dict.update(
             {